Move objects out of __init__
