+++ src/csrs/pages/templates/__init__.py
@@ -6,251 +6,17 @@ from jinja2 import Environment, Template
 
 from ... import schemas
 from ..loader import ENV
+from .edit_assumption import EditableAssumption
+from .edit_path import EditablePaths
+from .edit_run import EditableRuns
+from .edit_scenario import EditableScenario
 from .form_parts import (
     CreateSelection,
     CreateSelectionGroup,
     CreateStr,
     CreateStrLong,
     CreateSwitch,
-    EditableSelection,
-    EditableSelectionGroup,
-    EditableStr,
-    EditableStrLong,
-    EditableSwitch,
 )
-
-
-class EditableAssumption:
-    def __init__(
-        self,
-        obj: schemas.Assumption,
-        assumption_kinds: list[str],
-    ):
-        self.obj = obj
-        self.kinds = assumption_kinds
-        self.env: Environment = ENV
-
-    def render(self, request: Request, edit_on: bool = True) -> str:
-        # Pre render the editable sections
-        name = EditableStr(
-            id=self.obj.id,
-            name="name",
-            default=self.obj.name,
-        ).render(request)
-        kind = EditableSelection(
-            id=self.obj.id,
-            name="kind",
-            default=self.obj.kind,
-            options=self.kinds,
-        ).render(request)
-        detail = EditableStrLong(
-            id=self.obj.id,
-            name="detail",
-            default=self.obj.detail,
-            rows=max(1, self.obj.detail.count("\n")),
-        ).render(request)
-        # render the whole card
-        return self.env.get_template("templates/assumption.jinja").render(
-            edit_on=edit_on,
-            id=self.obj.id,
-            title=self.obj.name,
-            name=name,
-            kind=kind,
-            detail=detail,
-            request=request,
-        )
-
-
-class EditableScenario:
-    def __init__(
-        self,
-        obj: schemas.Scenario,
-        scenario_versions: list[str],
-        assumptions: list[schemas.Assumption],
-    ):
-        self.obj = obj
-        self.versions = scenario_versions
-        self.env: Environment = ENV
-        self.assumptions: dict[str, list[schemas.Assumption]] = dict()
-        for a in assumptions:
-            if a.kind not in self.assumptions:
-                self.assumptions[a.kind] = list()
-            self.assumptions[a.kind].append(a)
-
-    def render(self, request: Request, edit_on: bool = True) -> str:
-        # Pre render the editable sections
-        name = EditableStr(
-            id=self.obj.id,
-            name="name",
-            default=self.obj.name,
-        ).render(
-            request,
-            name_col_width=2,
-        )
-        preferred_run = EditableSelection(
-            id=self.obj.id,
-            name="preferred_run",
-            default=self.obj.preferred_run,
-            options=self.versions,
-        ).render(
-            request,
-            name_col_width=2,
-        )
-        assumptions_dropdowns = list()
-        for kind, assumption_name in self.obj.assumptions.items():
-            try:
-                options = list(a.name for a in self.assumptions[kind])
-            except KeyError:
-                options = list()
-            assumptions_dropdowns.append(
-                EditableSelection(
-                    id=self.obj.id,
-                    name=kind,
-                    default=assumption_name,
-                    options=options,
-                )
-            )
-        assumptions = EditableSelectionGroup(
-            name="assumptions",
-            editable_selections=assumptions_dropdowns,
-        ).render(request)
-
-        # render the whole card
-        return self.env.get_template("templates/scenario.jinja").render(
-            edit_on=edit_on,
-            request=request,
-            id=self.obj.id,
-            title=self.obj.name,
-            name=name,
-            preferred_run=preferred_run,
-            assumptions=assumptions,
-        )
-
-
-class EditableRuns:
-    def __init__(
-        self,
-        obj: schemas.Run,
-    ):
-        self.obj = obj
-        self.env: Environment = ENV
-
-    def render(self, request: Request, edit_on: bool = True) -> str:
-        _id = self.obj.id
-        scenario = EditableStr(
-            id=_id,
-            name="scenario",
-            default=self.obj.scenario,
-        ).render(
-            request,
-            toggle_tag="",
-        )
-        version = EditableStr(
-            id=_id,
-            name="version",
-            default=self.obj.version,
-        ).render(
-            request,
-            toggle_tag="",
-        )
-        contact = EditableStr(
-            id=_id,
-            name="contact",
-            default=self.obj.contact,
-        ).render(request)
-        confidential = EditableSwitch(
-            id=_id,
-            name="confidential",
-            default=self.obj.confidential,
-        ).render(request)
-        published = EditableSwitch(
-            id=_id,
-            name="published",
-            default=self.obj.published,
-        ).render(request)
-        code_version = EditableStr(
-            id=_id,
-            name="code_version",
-            default=self.obj.code_version,
-        ).render(request)
-        detail = EditableStrLong(
-            id=_id,
-            name="detail",
-            default=self.obj.detail,
-        ).render(request)
-        return self.env.get_template("templates/run.jinja").render(
-            edit_on=edit_on,
-            request=request,
-            title=f"{self.obj.scenario} <code>(v{self.obj.version})</code>",
-            scenario=scenario,
-            version=version,
-            contact=contact,
-            confidential=confidential,
-            published=published,
-            code_version=code_version,
-            detail=detail,
-            id=self.obj.id,
-        )
-
-
-class EditablePaths:
-    def __init__(
-        self,
-        obj: schemas.NamedPath,
-    ):
-        self.obj = obj
-        self.env: Environment = ENV
-
-    def render(self, request: Request, edit_on: bool = True) -> str:
-        _id = self.obj.id
-        name = EditableStr(
-            id=_id,
-            name="name",
-            default=self.obj.name,
-        ).render(request)
-        path = EditableStr(
-            id=_id,
-            name="path",
-            default=self.obj.path,
-        ).render(request)
-        category = EditableStr(
-            id=_id,
-            name="category",
-            default=self.obj.category,
-        ).render(request)
-        detail = EditableStrLong(
-            id=_id,
-            name="detail",
-            default=self.obj.detail,
-        ).render(request)
-        period_type = EditableStr(
-            id=_id,
-            name="period_type",
-            default=self.obj.period_type,
-        ).render(request)
-        interval = EditableStr(
-            id=_id,
-            name="interval",
-            default=self.obj.interval,
-        ).render(request)
-        units = EditableStr(
-            id=_id,
-            name="units",
-            default=self.obj.units,
-        ).render(request)
-        return self.env.get_template("templates/path.jinja").render(
-            edit_on=edit_on,
-            request=request,
-            title=self.obj.name,
-            name=name,
-            path=path,
-            category=category,
-            detail=detail,
-            period_type=period_type,
-            interval=interval,
-            units=units,
-            id=self.obj.id,
-        )
 
 
 class NewAssumption:
