+++ src/csrs/pages/templates/edit_path.jinja
