+++ src/csrs/pages/templates/edit_scenario.jinja
